refactor: rename file to fit its function name
